requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
mysql-connector-python>=8.3

chromadb>=0.5
sentence-transformers>=2.6

openai>=1.30

pandas>=2.0
langchain-text-splitters>=0.2